backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==1.10.12
pyjwt==2.8.0
python-multipart==0.0.6
